+++ resources/views/cart/show.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('My Cart')">
+    <livewire:storefront.cart-page />
+</x-layouts::storefront>